Finalização do processo de cadastro dos locais de votação

+++ src/main/java/br/jus/tse/distribuicao_urnas/csv/LocalVotacaoCSVDto.java
@@ -14,7 +14,7 @@ public class LocalVotacaoCSVDto {
 	@CsvBindByName
 	private Long numeroZE;
 	@CsvBindByName
-	private Integer numeroLocalVotacao;
+	private Long numeroLocalVotacao;
 	@CsvBindByName
 	private String nome;
 	@CsvBindByName
